Refactor reduction process tests and enhance output validation; improve STAC API reader with dimension estimation and error handling
